moving enum type to its own file
