+++ packages/app-desktop/gui/NoteEditor/NoteBody/TinyMCE/utils/types.ts
@@ -1,0 +1,15 @@
+// eslint-disable-next-line import/prefer-default-export
+export enum TinyMceEditorEvents {
+	KeyUp = 'keyup',
+	KeyDown = 'keydown',
+	KeyPress = 'keypress',
+	Paste = 'paste',
+	PasteAsText = 'pasteAsText',
+	Copy = 'copy',
+	CompositionEnd = 'compositionend',
+	Cut = 'cut',
+	JoplinChange = 'joplinChange',
+	Undo = 'Undo',
+	Redo = 'Redo',
+	ExecCommand = 'ExecCommand',
+}
